+++ crates/db/src/utils.rs
@@ -1,4 +1,5 @@
-pub const TEMP_PLACEHOLDER_NUM_TABLES: usize = 2;
+use crate::kv::table::{Decode, Table};
+use std::borrow::Cow;
 
 /// Returns the default page size that can be used in this OS.
 pub fn default_page_size() -> usize {
@@ -13,4 +14,29 @@ pub fn default_page_size() -> usize {
     } else {
         os_page_size
     }
+}
+
+/// Helper function to decode a `(key, value)` pair.
+pub fn decoder<'a, T>(kv: (Cow<'a, [u8]>, Cow<'a, [u8]>)) -> eyre::Result<(T::Key, T::Value)>
+where
+    T: Table,
+    T::Key: Decode,
+{
+    Ok((Decode::decode(&kv.0)?, Decode::decode(&kv.1)?))
+}
+
+/// Helper function to decode only a value from a `(key, value)` pair.
+pub fn decode_value<'a, T>(kv: (Cow<'a, [u8]>, Cow<'a, [u8]>)) -> eyre::Result<T::Value>
+where
+    T: Table,
+{
+    Decode::decode(&kv.1)
+}
+
+/// Helper function to decode a value. It can be a key or subkey.
+pub fn decode_one<'a, T>(value: Cow<'a, [u8]>) -> eyre::Result<T::Value>
+where
+    T: Table,
+{
+    Decode::decode(&value)
 }
